requirements.txt
astroid==2.2.5
cffi==1.12.2
Click==7.0
colorama==0.4.1
dnspython==1.16.0
Flask==1.0.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
pycparser==2.19
pylint==2.3.1
pymongo==3.7.2
python-engineio==3.4.3
python-socketio==4.0.0
six==1.12.0
typed-ast==1.3.1
Werkzeug==0.14.1
wrapt==1.11.1
